requirements.txt
mcp[cli]
python-dotenv
pdfminer.six
fastmcp
PyMuPDF
pytesseract
pillow